requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
psycopg2-binary>=2.9 # PostgreSQL adapter
python-dotenv>=0.20
Flask-Migrate>=4.0 # For database migrations
requests>=2.28      # For potential scraping later
beautifulsoup4>=4.11 # For potential scraping later
# Add other libraries like gunicorn for production later